test(engines): move test_leftovers back to the eth_engine file and make function private again
